# src/api/openapi.yaml
openapi: 3.1.0
info:
  title: Cashlint API
  version: 1.0.0
  description: Accrual-based Accounting App
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    IncomeStatement:
      type: object
      properties:
        period:
          type: object
          properties:
            startDate:
              type: string
              format: date-time
            endDate:
              type: string
              format: date-time
          required:
            - startDate
            - endDate
        revenues:
          type: array
          items:
            type: object
            properties:
              accountCode:
                type: string
              accountName:
                type: string
              amount:
                type: number
                format: decimal
            required:
              - accountCode
              - accountName
              - amount
        expenses:
          type: array
          items:
            type: object
            properties:
              accountCode:
                type: string
              accountName:
                type: string
              amount:
                type: number
                format: decimal
            required:
              - accountCode
              - accountName
              - amount
        revenueTotal:
          type: number
          format: decimal
        expenseTotal:
          type: number
          format: decimal
        netIncome:
          type: number
          format: decimal
      required:
        - period
        - revenues
        - expenses
        - revenueTotal
        - expenseTotal
        - netIncome
    BalanceSheet:
      type: object
      properties:
        asOfDate:
          type: string
          format: date-time
        assets:
          type: array
          items:
            type: object
            properties:
              accountCode:
                type: string
              accountName:
                type: string
              amount:
                type: number
                format: decimal
            required:
              - accountCode
              - accountName
              - amount
        liabilities:
          type: array
          items:
            type: object
            properties:
              accountCode:
                type: string
              accountName:
                type: string
              amount:
                type: number
                format: decimal
            required:
              - accountCode
              - accountName
              - amount
        equity:
          type: array
          items:
            type: object
            properties:
              accountCode:
                type: string
              accountName:
                type: string
              amount:
                type: number
                format: decimal
            required:
              - accountCode
              - accountName
              - amount
        assetTotal:
          type: number
          format: decimal
        liabilityTotal:
          type: number
          format: decimal
        equityTotal:
          type: number
          format: decimal
      required:
        - asOfDate
        - assets
        - liabilities
        - equity
        - assetTotal
        - liabilityTotal
        - equityTotal
    StatementOfOwnersEquity:
      type: object
      properties:
        period:
          type: object
          properties:
            startDate:
              type: string
              format: date-time
            endDate:
              type: string
              format: date-time
          required:
            - startDate
            - endDate
        beginningCapital:
          type: number
          format: decimal
        additionalContributions:
          type: number
          format: decimal
        netIncome:
          type: number
          format: decimal
        drawings:
          type: number
          format: decimal
        endingCapital:
          type: number
          format: decimal
      required:
        - period
        - beginningCapital
        - additionalContributions
        - netIncome
        - drawings
        - endingCapital
    StatementOfCashFlows:
      type: object
      properties:
        period:
          type: object
          properties:
            startDate:
              type: string
              format: date-time
            endDate:
              type: string
              format: date-time
          required:
            - startDate
            - endDate
        operatingActivities:
          type: array
          items:
            type: object
            properties:
              description:
                type: string
              amount:
                type: number
                format: decimal
            required:
              - description
              - amount
        investingActivities:
          type: array
          items:
            type: object
            properties:
              description:
                type: string
              amount:
                type: number
                format: decimal
            required:
              - description
              - amount
        financingActivities:
          type: array
          items:
            type: object
            properties:
              description:
                type: string
              amount:
                type: number
                format: decimal
            required:
              - description
              - amount
        netCashChange:
          type: number
          format: decimal
        beginningCash:
          type: number
          format: decimal
        endingCash:
          type: number
          format: decimal
      required:
        - period
        - operatingActivities
        - investingActivities
        - financingActivities
        - netCashChange
        - beginningCash
        - endingCash

    # Common Response Schemas
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            type:
              type: string
              enum: [DomainFailure, InfrastructureFailure, ApplicationFailure]
            subtype:
              type: string
            message:
              type: string
          required:
            - type
            - subtype
            - message
...
security:
  - cookieAuth: []
paths:
...
  /api/reporting/income-statement:
    get:
      summary: Generate income statement
      description: Generate an income statement for a given period.
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Income statement generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  incomeStatement:
                    $ref: '#/components/schemas/IncomeStatement'
                required:
                  - incomeStatement
        '400':
          $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/schemas/ErrorResponse'
  /api/reporting/balance-sheet:
    get:
      summary: Generate balance sheet
      description: Generate a balance sheet as of a specific date.
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: asOfDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Balance sheet generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  balanceSheet:
                    $ref: '#/components/schemas/BalanceSheet'
                required:
                  - balanceSheet
        '400':
          $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/schemas/ErrorResponse'
  /api/reporting/owners-equity:
    get:
      summary: Generate statement of owner's equity
      description: Generate a statement of owner's equity for a given period.
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Statement generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  statementOfOwnersEquity:
                    $ref: '#/components/schemas/StatementOfOwnersEquity'
                required:
                  - statementOfOwnersEquity
        '400':
          $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/schemas/ErrorResponse'
  /api/reporting/cash-flow:
    get:
      summary: Generate statement of cash flows
      description: Generate a statement of cash flows for a given period.
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Statement generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  statementOfCashFlows:
                    $ref: '#/components/schemas/StatementOfCashFlows'
                required:
                  - statementOfCashFlows
        '400':
          $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/schemas/ErrorResponse'

  # Health endpoints for new contexts
...